Add files via upload
K-means is a clustering algorithm used in unsupervised machine learning that partitions data according to stochastically generated centroids. Say 3 centers are generated amongst a group of 10 data points: data points A through C are closest to center A and as such, call out to it, drawing it closer; data points D through F are closest to center B and call out as well, the remaining data points and center adhering to the same trend. A cluster center inches closer to its data points by finding the mean of the distances between itself and said points. Outliers have a higher chance of switching clusters as the centers move to accommodate those closest to them.

This algorithm was made as a part of a bigger project within my undergraduate research in which I handled swarm robotics.

Any and all criticism is welcomed.